vtype: adding VLong and VLongArray factory implementations

+++ epics-vtype/src/main/java/org/epics/vtype/VLong.java
@@ -16,5 +16,5 @@ public interface VLong extends VNumber, VType {
      * {@inheritDoc }
      */
     @Override
-    Integer getValue();
+    Long getValue();
 }
